+++ app/page.tsx
@@ -1,19 +1,16 @@
 import { Footer } from '#/ui/Footer';
 import { Header } from '#/ui/Header';
 import { HomePage } from '#/ui/HomePage';
-import { fetchData } from '#/lib/api.js';
 
 import '#/css/print.css';
 import '#/css/index.css';
 
 export default async function Page() {
-  const data = await fetchData();
-
   return (
     <div>
       <Header />
 
-      <HomePage data={data} />
+      <HomePage />
 
       <Footer />
     </div>
