+++ ui/ProjectModal.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Modal from 'react-bootstrap/Modal';
 import Button from 'react-bootstrap/Button';
 import { sortBy } from 'lodash';
